Refactor onSelectExperimentsChange
